Rename existing form action
